Added lexer hooks for metadata_[key|value]
This hook will run once one on the first invocation of the code block,
and if it identifies an hr block ('---') as the first block it will
assume that we are handling metadata.

Metadata is a key/value store in the top of the document, when present
it begins with '---' and ends with '---' and is formated "<key>: <value>\n"
between.

Assume the following document:

    ---
    title: This is a title
    author: The Doctor
    ---

    Hello World

This token stream would begin with 'hr', contain word, space and new
line definitions to comprise the key/value lines, and end with an hr.

This is compressed into metadata_key token, followed by a metadata_value
token before being passed to the parser and the origional tokens are
forgotten.